+++ docs/changes.md
@@ -2,7 +2,8 @@
 
 * 2020/07/22
   * 增加 OBJECT\_CMD\_REMOVE\_CHECKED 、OBJECT\_PROP\_CHECKED和 OBJECT\_PROP\_SIZE。
-  
+  * 完善 value\_double，支持bool到double转换。
+
 * 2020/07/21
   * 完善文档。
   * 完善 API 文档增加导出函数。